requirements.txt
asgiref==3.5.2
asttokens==2.0.7
attrs==22.1.0
backcall==0.2.0
black==22.6.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
executing==0.9.1
factory-boy==3.2.1
Faker==15.3.4
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
model-bakery==1.9.0
mypy-extensions==0.4.3
packaging==22.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.12.0
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
tzdata==2022.7
wcwidth==0.2.5
